+++ src/components-v1/credentials-form.jsx
@@ -1,61 +1,49 @@
 import React, { useCallback, useState, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { find } from 'lodash';
 import { Button, Input, PasswordInput, Row, Text } from '@zextras/zapp-ui';
-import { postV1Login } from '../services/v1-service';
+
+const urlParams = new URLSearchParams(window.location.search);
 
 export default function CredentialsForm({
-	showAuthError,
-	handleSubmitCredentialsResponse,
+	authError,
+	submitCredentials,
 	configuration,
-	disableInputs
+	disableInputs,
+	loading = false
 }) {
-	const [ t ] = useTranslation();
+	const [t] = useTranslation();
 
-	const [ username, setUsername ] = useState('');
-	const [ password, setPassword ] = useState('');
+	const [username, setUsername] = useState(urlParams.get('username') || '');
+	const [password, setPassword] = useState('');
 
-	const submitUserPassword = useCallback((ev) => {
-		ev.preventDefault();
-
-		if (!username || !password) return;
-
-		postV1Login(
-			'PASSWORD',
-			username,
-			password
-		).then((res) => {
-			if (res.status === 200) {
-				if (window.PasswordCredential) {
-					// eslint-disable-next-line no-undef
-					const cred = new PasswordCredential({
-						id: username,
-						password,
-						name: password
-					});
-					navigator.credentials.store(cred);
-				}
-				window.location.assign(configuration.destinationUrl);
+	const submitUserPassword = useCallback((e) => {
+		e.preventDefault();
+		if (username && password) {
+			let usernameModified = username;
+			if (urlParams.has('virtualacctdomain')) {
+				usernameModified = `${usernameModified.replace('@', '.')}@${urlParams.get('virtualacctdomain')}`;
 			}
-			else {
-				handleSubmitCredentialsResponse(res);
+			else if (urlParams.has('customerDomain')){
+				usernameModified = `${usernameModified.trim()}@${urlParams.get('customerDomain')}`;
 			}
-		});
-	}, [username, password, configuration.destinationUrl, handleSubmitCredentialsResponse]);
+			submitCredentials(usernameModified, password);
+		}
+	}, [username, password, submitCredentials]);
 
 	const samlButtonCbk = useCallback(() => {
 		window.location.assign(`/zx/auth/startSamlWorkflow?redirectUrl=${configuration.destinationUrl}`);
 	}, [configuration]);
 
 	const samlButton = useMemo(() => {
-		if (find(configuration.authMethods, 'saml')) {
+		if (configuration.authMethods.includes('saml')) {
 			return (
 				<Button
 					type="outlined"
 					label={t('login_saml', 'Login SAML')}
 					color="primary"
 					disabled={disableInputs}
-					onClick={samlButtonCbk}/>
+					onClick={samlButtonCbk}
+				/>
 			);
 		}
 
@@ -65,14 +53,18 @@ export default function CredentialsForm({
 		);
 	}, [configuration, disableInputs, samlButtonCbk, t]);
 
+	const onChangeUsername = useCallback((ev) => setUsername(ev.target.value), [setUsername]);
+	const onChangePassword = useCallback((ev) => setPassword(ev.target.value), [setPassword]);
+
 	return (
-		<form style={{ width: '100%' }}>
+		<form onSubmit={submitUserPassword} style={{ width: '100%' }}>
+			<input type="submit" style={{ display: 'none' }}/>
 			<Row padding={{ bottom: 'large' }}>
 				<Input
-					value={username}
+					defaultValue={username}
 					disabled={disableInputs}
-					onChange={(ev) => setUsername(ev.target.value)}
-					hasError={showAuthError}
+					onChange={onChangeUsername}
+					hasError={!!authError}
 					autocomplete="username"
 					label={t('username','Username')}
 					backgroundColor="gray5"
@@ -80,21 +72,20 @@ export default function CredentialsForm({
 			</Row>
 			<Row padding={{ bottom: 'small' }}>
 				<PasswordInput
-					value={password}
+					defaultValue={password}
 					disabled={disableInputs}
-					onChange={(ev) => setPassword(ev.target.value)}
-					hasError={showAuthError}
+					onChange={onChangePassword}
+					hasError={!!authError}
 					autocomplete="password"
 					label={t('password', 'Password')}
 					backgroundColor="gray5"
 				/>
 			</Row>
 			<Text color="error" size="medium" overflow="break-word">
-				{showAuthError && t('credentials_not_valid','Credentials are not valid, please check data and try again')}
-				{!showAuthError && <br/>}
+				{authError || <br/>}
 			</Text>
 			<Row orientation="vertical" crossAlignment="flex-start" padding={{ bottom: 'large', top: 'small' }}>
-				<Button onClick={submitUserPassword} disabled={disableInputs} label={t('login','Login')} size="fill"/>
+				<Button loading={loading} onClick={submitUserPassword} disabled={disableInputs} label={t('login','Login')} size="fill" />
 			</Row>
 			<Row mainAlignment="flex-end" padding={{ bottom: 'extralarge' }}>
 				{samlButton}