+++ src/components-v1/form-selector.jsx
@@ -1,26 +1,24 @@
 import React, { useEffect, useState } from 'react';
 import V1LoginManager from './v1-login-manager';
-import { getAuthSupported } from '../services/auth-configuration-service';
+import V2LoginManager from './v2-login-manager';
+import { getAuthSupported, doAuthLogout } from '../services/auth-configuration-service';
 import ServerNotResponding from '../components-index/server-not-responding';
 import NotSupportedVersion from '../components-index/not-supported-version';
 
-export default function FormSelector ({ publicUrl }) {
+export default function FormSelector({ destinationUrl, domain }) {
 	const [ configuration, setConfiguration ] = useState(null);
 	const [ error, setError ] = useState(false);
 	const [ disableInputs, setDisableInputs ] = useState(true);
 
 	useEffect(() => {
 		let componentIsMounted = true;
-		const urlParams = new URLSearchParams(window.location.search);
-		const domain = urlParams.get('domain');
-		const destinationUrl = urlParams.get('destinationUrl');
 
 		getAuthSupported(domain)
 			.then((res) => {
 				if (componentIsMounted) {
 					setConfiguration({
 						...res,
-						destinationUrl: destinationUrl || publicUrl
+						destinationUrl
 					});
 					setDisableInputs(false);
 				}
@@ -32,7 +30,15 @@ export default function FormSelector ({ publicUrl }) {
 		return () => {
 			componentIsMounted = false;
 		}
-	}, [publicUrl]);
+	}, []);
+
+	useEffect(() => {
+		const urlParams = new URLSearchParams(window.location.search);
+		if (configuration && urlParams.has('loginOp') && urlParams.get('loginOp') === 'logout') {
+			doAuthLogout(configuration)
+				.catch(() => console.log('Logout failed'));
+		}
+	}, [configuration]);
 
 	if(error)
 		return <ServerNotResponding />
@@ -40,7 +46,12 @@ export default function FormSelector ({ publicUrl }) {
 	if (configuration === null)
 		return <div></div>;
 
-	if (configuration.maxApiVersion >= 1  &&  configuration.minApiVersion <= 1)
+	if (configuration.maxApiVersion >= 2 && configuration.minApiVersion <= 2)
+		return <V2LoginManager
+			configuration={configuration}
+			disableInputs={disableInputs}
+		/>;
+	if (configuration.maxApiVersion >= 1 && configuration.minApiVersion <= 1)
 		return <V1LoginManager
 			configuration={configuration}
 			disableInputs={disableInputs}