+++ src/mocks/auth/v1/logout.js
@@ -1,0 +1,10 @@
+import { rest } from 'msw';
+
+export default rest.get(
+	'/zx/auth/v1/logout',
+	(req, res, ctx) => {
+		return res(
+			ctx.status(200)
+		);
+	}
+);