+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.23-beta.2](https://bitbucket.org/zextras/zapp-login/compare/v0.0.23-beta.1...v0.0.23-beta.2) (2021-03-29)
+
+### [0.0.23-beta.1](https://bitbucket.org/zextras/zapp-login/compare/v0.0.23-beta.0...v0.0.23-beta.1) (2021-03-20)
+
+
+### Features
+
+* managed the loginPageTitle attribute ([d949d56](https://bitbucket.org/zextras/zapp-login/commit/d949d569e4c87b5b79c31c22c04b7d27bb94424a))
+
+
+### Bug Fixes
+
+* fixed some issues with the OTP authentication and updated the DS ([4a8b964](https://bitbucket.org/zextras/zapp-login/commit/4a8b964b5b9c3a5631da66dd86ca9439552fc5b6))
+* fixed the device-uuid generation ([3a360e4](https://bitbucket.org/zextras/zapp-login/commit/3a360e482cf39e8756561cd2a40b640d9d48e9c6))
+* renamed loginPageColorPalette to loginPageColorSet and fixed custom color assignment ([670a193](https://bitbucket.org/zextras/zapp-login/commit/670a193599b9cbe9e4e60aa008ba8fbec630a2aa))
+
+### [0.0.23-beta.0](https://bitbucket.org/zextras/zapp-login/compare/v0.0.22...v0.0.23-beta.0) (2021-01-12)
+
+
+### Features
+
+* added otp selector ([2df3090](https://bitbucket.org/zextras/zapp-login/commit/2df30904474dcd12cafd4c27f7239d45dd110104))
+* added two-factor authentication (v2-service) ([3395888](https://bitbucket.org/zextras/zapp-login/commit/33958887adf472cb44428a1167608473d1dae90b))
+
+
+### Bug Fixes
+
+* due to mockup release ([270ed82](https://bitbucket.org/zextras/zapp-login/commit/270ed8258a5b9dc692858cce5d60506e3e998db1))
+* not passing 'domain' parameter if domain is missing ([b59f8d2](https://bitbucket.org/zextras/zapp-login/commit/b59f8d214a66a1f7480251b94a10c9f753756ed9))
+
 ### [0.0.22](https://bitbucket.org/zextras/zapp-login/compare/v0.0.22-beta.0...v0.0.22) (2020-12-17)
 
 ### [0.0.22-beta.0](https://bitbucket.org/zextras/zapp-login/compare/v0.0.21...v0.0.22-beta.0) (2020-12-17)